Feat : Add Price Oracle functionality        Add deploy script for underlying token and price oracle        Added comptroller

+++ contracts/BaseInterestRate.sol
@@ -1,7 +1,7 @@
 // // SPDX-License-Identifier: Unlicense
 pragma solidity 0.8.19;
 
-import "./InterestRateInterface.sol";
+import "./Interfaces/InterestRateModelInterface.sol";
 
 contract InterestRateModel is InterestRateModelInterface {
     event InterestRateUpdated(
@@ -21,27 +21,27 @@ contract InterestRateModel is InterestRateModelInterface {
     /**
      * @notice The approximate number of blocks per year that is assumed by the interest rate model
      */
-    uint public constant blocksPerYear = 2102400;
+    uint256 public constant blocksPerYear = 2102400;
 
     /**
      * @notice The multiplier of utilization rate that gives the slope of the interest rate
      */
-    uint public multiplierPerBlock;
+    uint256 public multiplierPerBlock;
 
     /**
      * @notice The base interest rate which is the y-intercept when utilization rate is 0
      */
-    uint public baseRatePerBlock;
+    uint256 public baseRatePerBlock;
 
     /**
      * @notice The multiplierPerBlock after hitting a specified utilization point
      */
-    uint public jumpMultiplierPerBlock;
+    uint256 public jumpMultiplierPerBlock;
 
     /**
      * @notice The utilization point at which the jump multiplier is applied
      */
-    uint public kink;
+    uint256 public kink;
 
     /**
      * @notice Construct an interest rate model
@@ -52,10 +52,10 @@ contract InterestRateModel is InterestRateModelInterface {
      * @param owner_ The address of the owner, i.e. the Timelock contract (which has the ability to update parameters directly)
      */
     constructor(
-        uint baseRatePerYear,
-        uint multiplierPerYear,
-        uint jumpMultiplierPerYear,
-        uint kink_,
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink_,
         address owner_
     ) {
         owner = owner_;
@@ -76,10 +76,10 @@ contract InterestRateModel is InterestRateModelInterface {
      * @param kink_ The utilization point at which the jump multiplier is applied
      */
     function updateJumpRateModel(
-        uint baseRatePerYear,
-        uint multiplierPerYear,
-        uint jumpMultiplierPerYear,
-        uint kink_
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink_
     ) external virtual {
         require(msg.sender == owner, "only the owner may call this function.");
 
@@ -99,10 +99,10 @@ contract InterestRateModel is InterestRateModelInterface {
      * @return The utilization rate as a mantissa between [0, BASE]
      */
     function utilizationRate(
-        uint cash,
-        uint borrows,
-        uint reserves
-    ) public pure returns (uint) {
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves
+    ) public pure returns (uint256) {
         if (borrows == 0) {
             return 0;
         }
@@ -118,18 +118,18 @@ contract InterestRateModel is InterestRateModelInterface {
      * @return The borrow rate percentage per block as a mantissa (scaled by BASE)
      */
     function getBorrowRate(
-        uint cash,
-        uint borrows,
-        uint reserves
-    ) public view returns (uint) {
-        uint util = utilizationRate(cash, borrows, reserves);
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves
+    ) public view returns (uint256) {
+        uint256 util = utilizationRate(cash, borrows, reserves);
 
         if (util <= kink) {
             return ((util * multiplierPerBlock) / BASE) + baseRatePerBlock;
         } else {
-            uint normalRate = ((kink * multiplierPerBlock) / BASE) +
+            uint256 normalRate = ((kink * multiplierPerBlock) / BASE) +
                 baseRatePerBlock;
-            uint excessUtil = util - kink;
+            uint256 excessUtil = util - kink;
             return ((excessUtil * jumpMultiplierPerBlock) / BASE) + normalRate;
         }
     }
@@ -143,14 +143,14 @@ contract InterestRateModel is InterestRateModelInterface {
      * @return The supply rate percentage per block as a mantissa (scaled by BASE)
      */
     function getSupplyRate(
-        uint cash,
-        uint borrows,
-        uint reserves,
-        uint reserveFactorMantissa
-    ) public view virtual returns (uint) {
-        uint oneMinusReserveFactor = BASE - reserveFactorMantissa;
-        uint borrowRate = getBorrowRate(cash, borrows, reserves);
-        uint rateToPool = (borrowRate * oneMinusReserveFactor) / BASE;
+        uint256 cash,
+        uint256 borrows,
+        uint256 reserves,
+        uint256 reserveFactorMantissa
+    ) public view virtual returns (uint256) {
+        uint256 oneMinusReserveFactor = BASE - reserveFactorMantissa;
+        uint256 borrowRate = getBorrowRate(cash, borrows, reserves);
+        uint256 rateToPool = (borrowRate * oneMinusReserveFactor) / BASE;
         return (utilizationRate(cash, borrows, reserves) * rateToPool) / BASE;
     }
 
@@ -162,15 +162,13 @@ contract InterestRateModel is InterestRateModelInterface {
      * @param kink_ The utilization point at which the jump multiplier is applied
      */
     function updateJumpRateModelInternal(
-        uint baseRatePerYear,
-        uint multiplierPerYear,
-        uint jumpMultiplierPerYear,
-        uint kink_
+        uint256 baseRatePerYear,
+        uint256 multiplierPerYear,
+        uint256 jumpMultiplierPerYear,
+        uint256 kink_
     ) internal {
         baseRatePerBlock = baseRatePerYear / blocksPerYear;
-        multiplierPerBlock =
-            (multiplierPerYear * BASE) /
-            (blocksPerYear * kink_);
+        multiplierPerBlock = (multiplierPerYear * BASE) / (blocksPerYear * kink_);
         jumpMultiplierPerBlock = jumpMultiplierPerYear / blocksPerYear;
         kink = kink_;
